+++ src/TestSuite/TestCase.php
@@ -16,6 +16,7 @@ use Cake\Core\Configure;
 use Cake\Filesystem\Folder;
 use Cake\TestSuite\TestCase as CakeTestCase;
 use Thumber\ThumbsPathTrait;
+use Thumber\Utility\ThumbCreator;
 use Tools\ReflectionTrait;
 use Tools\TestSuite\TestCaseTrait;
 
@@ -93,7 +94,7 @@ abstract class TestCase extends CakeTestCase
      */
     public function assertThumbPath($path, $message = '')
     {
-        $regex = sprintf('/^%s[\w\d]{32}_[\w\d]{32}\.\w{3,4}/', preg_quote($this->getPath() . DS, '/'));
+        $regex = sprintf('/^%s[\w\d_]+\.\w{3,4}/', preg_quote($this->getPath() . DS, '/'));
         self::assertRegExp($regex, $path, $message);
     }
 
@@ -108,6 +109,45 @@ abstract class TestCase extends CakeTestCase
     public function assertThumbUrl($url, $message = '')
     {
         self::assertRegExp('/^(http:\/\/localhost)?\/thumb\/[\w\d]+/', $url, $message);
+    }
+
+    /**
+     * Returns an instance of `ThumbCreator`
+     * @param string $path Path of the image from which to create the
+     *  thumbnail. It can be a relative path (to APP/webroot/img), a full path
+     *  or a remote url
+     * @return ThumbCreator
+     * @since 1.5.1
+     */
+    protected function getThumbCreatorInstance($path = null)
+    {
+        return new ThumbCreator($path ?: '400x400.jpg');
+    }
+
+    /**
+     * Returns an instance of `ThumbCreator`, after calling `resize()` and
+     *  `save()` methods.
+     *
+     * It can be called passing only the array of options as first argument.
+     * @param string $path Path of the image from which to create the
+     *  thumbnail. It can be a relative path (to APP/webroot/img), a full path
+     *  or a remote url
+     * @param array $options Options for saving
+     * @return ThumbCreator
+     * @since 1.5.1
+     * @uses getThumbCreatorInstance()
+     */
+    protected function getThumbCreatorInstanceWithSave($path = null, array $options = [])
+    {
+        if (is_array($path) && func_num_args() < 2) {
+            $options = $path;
+            $path = null;
+        }
+
+        $thumbCreator = $this->getThumbCreatorInstance($path);
+        $thumbCreator->resize(200)->save($options);
+
+        return $thumbCreator;
     }
 
     /**